review comment: heavily used trait in own source
